+++ examples/db_instance/main.tf
@@ -68,8 +68,9 @@ module "db_instance" {
   # Security group
   #####
 
-  security_group_name         = "tftest"
-  security_group_source_cidrs = ["127.0.0.1/32", "10.0.0.0/8"]
+  security_group_name                = "tftest"
+  allowed_cidrs                      = ["127.0.0.1/32", "10.0.0.0/8"]
+  manage_client_security_group_rules = false
   security_group_tags = {
     dbsgtags = "tftest"
   }