+++ examples/rds-cluster/main.tf
@@ -83,8 +83,8 @@ module "rds_cluster" {
   # Security group
   #####
 
-  security_group_name         = "tftest"
-  security_group_source_cidrs = ["127.0.0.1/32", "10.0.0.0/8"]
+  security_group_name = "tftest"
+  allowed_cidrs       = ["127.0.0.1/32", "10.0.0.0/8"]
   security_group_tags = {
     dbsgtags = "tftest"
   }