+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 2.1.0
 
-  * maitenance: bump SSM parameters module to 3.0.1
+  * maintenance: bump SSM parameters module to 3.0.1
   * fix: typo in versions.tf to be usable with terraform 0.13
 
 ## 2.0.0